Stop Duplication of Hydraulic File
Prior to changes in the msxtoolkit.c file, the CALL(err, ENsavehydfile(...)) would save a hydraulic file that was then referenced by MSX. They ENsolveH() had already created a binary file, so this change just uses the original binary. This saves the disk space when running, and also saves the time it takes to duplicate the file.